HBASE-22473 Split TestSCP
